requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
config==0.5.1
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Pillow==9.2.0
psycopg2-binary==2.9.4
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
django-filter==22.1
